+++ sdk/base/unittest/trace.hpp
@@ -48,6 +48,7 @@ enum trace_event_crypto_t {
     trace_event_cose_signing = 13,
     trace_event_cose_mac = 14,
     trace_event_keyexchange = 15,
+    trace_event_verify = 16,
 };
 
 enum trace_event_net_t {