+++ sdk/base/unittest/trace_advisor.cpp
@@ -50,6 +50,7 @@ void trace_advisor::load() {
                 e.event_map.insert({trace_event_cose_signing, "COSE signing"});
                 e.event_map.insert({trace_event_cose_mac, "COSE mac"});
                 e.event_map.insert({trace_event_keyexchange, "keyexchange"});
+                e.event_map.insert({trace_event_verify, "verify"});
                 _resource_map.insert({trace_category_crypto, e});
             }
             {